+++ h2o-algos/src/main/java/hex/deeplearning/Storage.java
@@ -19,11 +19,11 @@ public class Storage {
    * Abstract vector interface
    */
   public abstract interface Vector {
-    public abstract double get(int i);
-    public abstract void set(int i, double val);
-    public abstract void add(int i, double val);
+    public abstract float get(int i);
+    public abstract void set(int i, float val);
+    public abstract void add(int i, float val);
     public abstract int size();
-    public abstract double[] raw();
+    public abstract float[] raw();
     public abstract Frame toFrame(Key key);
   }
 
@@ -60,14 +60,14 @@ public class Storage {
    * Dense vector implementation
    */
   public static class DenseVector extends Iced implements Vector {
-    private double[] _data;
-    DenseVector(int len) { _data = new double[len]; }
-    DenseVector(double[] v) { _data = v; }
-    @Override public double get(int i) { return _data[i]; }
-    @Override public void set(int i, double val) { _data[i] = val; }
-    @Override public void add(int i, double val) { _data[i] += val; }
+    private float[] _data;
+    DenseVector(int len) { _data = new float[len]; }
+    DenseVector(float[] v) { _data = v; }
+    @Override public float get(int i) { return _data[i]; }
+    @Override public void set(int i, float val) { _data[i] = val; }
+    @Override public void add(int i, float val) { _data[i] += val; }
     @Override public int size() { return _data.length; }
-    @Override public double[] raw() { return _data; }
+    @Override public float[] raw() { return _data; }
     @Override public Frame toFrame(Key key) { return Storage.toFrame(this, key); }
   }
 
@@ -76,14 +76,14 @@ public class Storage {
    */
   public static class SparseVector extends Iced implements Vector {
     int[] _indices;
-    double[] _values;
+    float[] _values;
     private int _size;
     private int _nnz;
 
     @Override public int size() { return _size; }
     public int nnz() { return _nnz; }
 
-    SparseVector(double[] v) { this(new DenseVector(v)); }
+    SparseVector(float[] v) { this(new DenseVector(v)); }
     SparseVector(final DenseVector dv) {
       _size = dv.size();
       // first count non-zeros
@@ -94,7 +94,7 @@ public class Storage {
       }
       // only allocate what's needed
       _indices = new int[_nnz];
-      _values = new double[_nnz];
+      _values = new float[_nnz];
       // fill values
       int idx = 0;
       for (int i=0; i<dv._data.length; ++i) {
@@ -112,23 +112,23 @@ public class Storage {
      * @param i element index
      * @return real value
      */
-    @Override public double get(int i) {
+    @Override public float get(int i) {
       final int idx = Arrays.binarySearch(_indices, i);
       return idx < 0 ? 0f : _values[idx];
     }
 
     @Override
-    public void set(int i, double val) {
+    public void set(int i, float val) {
       throw new UnsupportedOperationException("setting values in a sparse vector is not implemented.");
     }
 
     @Override
-    public void add(int i, double val) {
+    public void add(int i, float val) {
       throw new UnsupportedOperationException("adding values in a sparse vector is not implemented.");
     }
 
     @Override
-    public double[] raw() {
+    public float[] raw() {
       throw new UnsupportedOperationException("raw access to the data in a sparse vector is not implemented.");
     }
 
@@ -152,9 +152,9 @@ public class Storage {
       public String toString() {
         return index() + " -> " + value();
       }
-      double value() { return _values[_idx]; }
+      float value() { return _values[_idx]; }
       int index() { return _indices[_idx]; }
-      void setValue(double val) { _values[_idx] = val; }
+      void setValue(float val) { _values[_idx] = val; }
     }
 
     public Iterator begin() { return new Iterator(0); }
